Made to compile by MSVC 2015/2017 32 bit compilers (#101)
Patches by @anatoly-spb :
1. Get rid of -DWIN64, -DZLIB_WINAPI in 32-bit build for fixing zlib linkage;
2. Work around MSVC2015 limitation based on lacking the C++14 features (like void as literal type) and document it.

+++ Telegram/SourceFiles/core/basic_types.h
@@ -54,4 +54,4 @@ void as_const(const T&&) = delete;
 #include <cstdint>
 
 #define qsl(s) QStringLiteral(s)
-#define qstr(s) QLatin1String(s, sizeof(s) - 1)
+#define qstr(s) QLatin1String(s, static_cast<int>(sizeof(s)-1))
